4chanalerter.py
This code retrieves 4chan.org/b. It then parses it with beautiful soup and prettifys it. It then searches for strings of your choice (comma delimited). If matches are found it will send an email to you from another email account. It will also archive the page where it found the match.